+++ Sensor.h
@@ -27,7 +27,7 @@ public:
   //report of current Sensor state(with name)
   virtual String report() const = 0;
   //get id of sensor
-  virtual unsigned int id() const = 0;
+  virtual unsigned long id() const = 0;
 protected:  
   char _name[SENSOR_NAME_LENGTH];
   SENSOR_T _type;
